# docs/api-spec.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryPerfRow": {
        "properties": {
          "changePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changepct"
          },
          "composite": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Composite"
          },
          "latestClose": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestclose"
          },
          "latestDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestdate"
          },
          "long": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Long"
          },
          "mid": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mid"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "short": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short"
          },
          "snapshotClose": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshotclose"
          },
          "stockId": {
            "title": "Stockid",
            "type": "string"
          },
          "stockName": {
            "title": "Stockname",
            "type": "string"
          },
          "strategies": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategies"
          }
        },
        "required": [
          "stockId",
          "stockName"
        ],
        "title": "HistoryPerfRow",
        "type": "object"
      },
      "HistoryPerfSummary": {
        "properties": {
          "asOf": {
            "title": "Asof",
            "type": "string"
          },
          "avgChangePct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avgchangepct"
          },
          "daysElapsed": {
            "default": 0,
            "title": "Dayselapsed",
            "type": "integer"
          },
          "hitCount": {
            "default": 0,
            "title": "Hitcount",
            "type": "integer"
          },
          "latestDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestdate"
          },
          "lossCount": {
            "default": 0,
            "title": "Losscount",
            "type": "integer"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HistoryPerfRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          },
          "winCount": {
            "default": 0,
            "title": "Wincount",
            "type": "integer"
          },
          "winRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winrate"
          }
        },
        "required": [
          "asOf"
        ],
        "title": "HistoryPerfSummary",
        "type": "object"
      },
...
      "IndustryMemberRow": {
        "properties": {
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "ret1D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret1D"
          },
          "ret20D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret20D"
          },
          "ret5D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret5D"
          },
          "stockId": {
            "title": "Stockid",
            "type": "string"
          },
          "stockName": {
            "title": "Stockname",
            "type": "string"
          }
        },
        "required": [
          "stockId",
          "stockName"
        ],
        "title": "IndustryMemberRow",
        "type": "object"
      },
      "IndustryRotationResponse": {
        "description": "`/api/market/industry-rotation` 的回傳殼，多帶一個資料截止日期。",
        "properties": {
          "asOf": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asof"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IndustryRotationRow"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "title": "IndustryRotationResponse",
        "type": "object"
      },
      "IndustryRotationRow": {
        "properties": {
          "heat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heat"
          },
          "industry": {
            "title": "Industry",
            "type": "string"
          },
          "nDown": {
            "default": 0,
            "title": "Ndown",
            "type": "integer"
          },
          "nFlat": {
            "default": 0,
            "title": "Nflat",
            "type": "integer"
          },
          "nMembers": {
            "default": 0,
            "title": "Nmembers",
            "type": "integer"
          },
          "nUp": {
            "default": 0,
            "title": "Nup",
            "type": "integer"
          },
          "ret1D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret1D"
          },
          "ret1DWeighted": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret1Dweighted"
          },
          "ret20D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret20D"
          },
          "ret5D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret5D"
          },
          "ret60D": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ret60D"
          },
          "totalAmount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Totalamount"
          }
        },
        "required": [
          "industry"
        ],
        "title": "IndustryRotationRow",
        "type": "object"
      },
...
      "RadarStrategy": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "hitCount": {
            "default": 0,
            "title": "Hitcount",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stocksOnly": {
            "default": false,
            "title": "Stocksonly",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "description"
        ],
        "title": "RadarStrategy",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Next.js 前端的 FastAPI 後端，包裝既有 app.* 模組。",
    "title": "台股研究儀表板 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/history/performance": {
      "get": {
        "description": "歷史追蹤 performance。\n- top=0 預設套用 hard cap (500 列)，避免一次回傳上千列拖慢前端\n- 排序：composite 降序（None 墊底）\n- 統計（勝率 / 平均漲幅）由完整集合計算，不受截斷影響\n- `summary.truncated`：被 hard cap 截掉時為 True，UI 應提示使用者用 strategy filter 縮小範圍\n- market filter：個股 (上市/上櫃) / ETF 分流",
        "operationId": "performance_api_history_performance_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "title": "As Of",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "strategy",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Strategy"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 500,
              "minimum": 0,
              "title": "Top",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "default": [
                "上市",
                "上櫃",
                "ETF"
              ],
              "items": {
                "type": "string"
              },
              "title": "Market",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPerfSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Performance",
        "tags": [
          "history"
        ]
      }
    },
    "/api/history/strategies": {
      "get": {
        "description": "指定 as_of 各策略命中數。可指定 market filter（個股 tab 排除 ETF）。\n當日命中數=0 的策略由前端決定要不要過濾。",
        "operationId": "strategies_for_date_api_history_strategies_get",
        "parameters": [
          {
            "in": "query",
            "name": "as_of",
            "required": true,
            "schema": {
              "title": "As Of",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "default": [
                "上市",
                "上櫃",
                "ETF"
              ],
              "items": {
                "type": "string"
              },
              "title": "Market",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RadarStrategy"
                  },
                  "title": "Response Strategies For Date Api History Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Strategies For Date",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/market/industry-members": {
      "get": {
        "operationId": "members_api_market_industry_members_get",
        "parameters": [
          {
            "in": "query",
            "name": "industry",
            "required": true,
            "schema": {
              "title": "Industry",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Top",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IndustryMemberRow"
                  },
                  "title": "Response Members Api Market Industry Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Members",
        "tags": [
          "market"
        ]
      }
    },
    "/api/market/industry-rotation": {
      "get": {
        "operationId": "rotation_api_market_industry_rotation_get",
        "parameters": [
          {
            "in": "query",
            "name": "min_members",
            "required": false,
            "schema": {
              "default": 3,
              "title": "Min Members",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndustryRotationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotation",
        "tags": [
          "market"
        ]
      }
    },
...
  }
}